fix: missing overrides, float/double return, uint8_t pair narrowing
Four warnings from the previous push:

- include/sw/kpu/timing/memory_controller_process.hpp:282
- include/sw/kpu/timing/dma_engine_process.hpp:190
  added `override` to is_complete - both override IProcess::is_complete
  (process_interface.hpp:232) and the class marks other overrides
  correctly, so clang -Winconsistent-missing-override fires.

- src/software/runtime/runtime.cpp:318
  restructured the elapsed_time arithmetic to do the divide in double
  and cast the final result to float. The previous form computed
  `static_cast<float>(delta) / (clock_ghz * 1e6f)` which is
  `float / double = double`, then implicitly narrowed back to float
  on return (MSVC C4244).

- src/software/dataflow/block_mover_compiler.cpp:575
  wrapped the two pair elements in static_cast<uint8_t>. uint8_t /
  uint8_t promotes to int via the usual arithmetic conversions; the
  std::pair<uint8_t,uint8_t> brace-init then narrows back (MSVC C4244
  fires from <utility> at the std::pair constructor instantiation).

Refs #23

+++ src/software/dataflow/block_mover_compiler.cpp
@@ -572,7 +572,11 @@ uint8_t BlockMoverCompiler::get_trigger_for_node(size_t node_id) {
 }
 
 std::pair<uint8_t, uint8_t> BlockMoverCompiler::get_mesh_position(uint8_t l3_id) const {
-    return {l3_id / config_.mesh_cols, l3_id % config_.mesh_cols};
+    // C integer promotion bumps uint8_t / uint8_t to int, then the pair
+    // brace-init narrows it back - MSVC C4244 in <utility>. Cast at the
+    // construction site so the narrowing is explicit.
+    return {static_cast<uint8_t>(l3_id / config_.mesh_cols),
+            static_cast<uint8_t>(l3_id % config_.mesh_cols)};
 }
 
 uint8_t BlockMoverCompiler::get_l3_id(uint8_t row, uint8_t col) const {
